refactor: ResultTest -> WinAndBonusTest, 정렬 추가
